Updated doc, fixed return values
