--- examples/model_selection/plot_learning_curve_times.py
@@ -1,127 +1,0 @@
-"""
-=====================================
-Plotting Learning Curve compute times
-=====================================
-
-On the left side, the training time of several learning curves are shown for
-various estimators. On the right side, the training time of those estimators
-are shown with their corresponding cross-validated score. Note that even if
-KNeighborsRegressor is the fastest estimator (as shown on the left side),
-SGDRegressor is the estimator giving the best score for this dataset (as shown
-on the right side).
-"""
-print(__doc__)
-
-import numpy as np
-import matplotlib.pyplot as plt
-from sklearn.datasets import make_regression
-from sklearn.linear_model import SGDRegressor
-from sklearn.svm import SVR
-from sklearn.neighbors import KNeighborsRegressor
-from sklearn.ensemble import RandomForestRegressor
-from sklearn.model_selection import ShuffleSplit, learning_curve
-
-
-def plot_learning_curve_times(estimators, X, y, train_sizes,
-                              cv=None, n_jobs=-1):
-    """
-    Generate a simple plot of the test and training learning curve.
-
-    Parameters
-    ----------
-    estimators : A list of object type that implement the "fit" and "predict"
-        methods. An object of that type which is cloned for each validation.
-
-    X : array-like, shape (n_samples, n_features)
-        Training vector, where n_samples is the number of samples and
-        n_features is the number of features.
-
-    y : array-like, shape (n_samples) or (n_samples, n_features), optional
-        Target relative to X for classification or regression;
-        None for unsupervised learning.
-
-    train_sizes : array-like, shape (n_ticks,), dtype float or int
-        Relative or absolute numbers of training examples that will be used to
-        generate the learning curve. If the dtype is float, it is regarded as a
-        fraction of the maximum size of the training set (that is determined
-        by the selected validation method), i.e. it has to be within (0, 1].
-        Otherwise it is interpreted as absolute sizes of the training sets.
-        Note that for classification the number of samples usually have to
-        be big enough to contain at least one sample from each class.
-
-    cv : int, cross-validation generator or an iterable, optional
-        Determines the cross-validation splitting strategy.
-        Possible inputs for cv are:
-          - None, to use the default 3-fold cross-validation,
-          - integer, to specify the number of folds.
-          - :term:`CV splitter`,
-          - An iterable yielding (train, test) splits as arrays of indices.
-
-        For integer/None inputs, if ``y`` is binary or multiclass,
-        :class:`StratifiedKFold` used. If the estimator is not a classifier
-        or if ``y`` is neither binary nor multiclass, :class:`KFold` is used.
-
-        Refer :ref:`User Guide <cross_validation>` for the various
-        cross-validators that can be used here.
-
-    n_jobs : int or None, optional (default=-1)
-        Number of jobs to run in parallel.
-        ``None`` means 1 unless in a :obj:`joblib.parallel_backend` context.
-        ``-1`` means using all processors. See :term:`Glossary <n_jobs>`
-        for more details.
-    """
-    fig, axes = plt.subplots(1, 2, figsize=(18, 6))
-
-    axes[0].set_title("Learning Curves computation times")
-    axes[0].set_xlabel("Training examples")
-    axes[0].set_ylabel("Training time (s)")
-    axes[0].grid()
-
-    axes[1].set_title("Estimators fit times with scores")
-    axes[1].set_xlabel("Fit times (s)")
-    axes[1].set_ylabel("Cross-validation score")
-    axes[1].grid()
-
-    for name, estimator in estimators:
-        train_sizes, _, test_scores, fit_times, _ = \
-            learning_curve(estimator(), X, y, cv=cv, n_jobs=n_jobs,
-                           train_sizes=train_sizes, return_times=True)
-
-        test_scores_mean = np.mean(test_scores, axis=1)
-        test_scores_std = np.std(test_scores, axis=1)
-        fit_times_mean = np.mean(fit_times, axis=1)
-        fit_times_std = np.std(fit_times, axis=1)
-
-        p = axes[0].plot(train_sizes, fit_times_mean, 'o-', label=name)
-
-        axes[0].fill_between(train_sizes, fit_times_mean - fit_times_std,
-                             fit_times_mean + fit_times_std, alpha=0.1,
-                             color=p[0].get_color())
-
-        p = axes[1].plot(fit_times_mean, test_scores_mean, 'o-',
-                         label=name)
-
-        axes[1].fill_between(fit_times_mean,
-                             test_scores_mean - test_scores_std,
-                             test_scores_mean + test_scores_std,
-                             alpha=0.1, color=p[0].get_color())
-
-    axes[0].legend(loc="best")
-    axes[1].legend(loc="best")
-    return fig
-
-
-X, y = make_regression(n_samples=int(1e4), n_features=50, n_informative=25,
-                       bias=-92, noise=100)
-
-cv = ShuffleSplit(n_splits=5, test_size=0.2, random_state=0)
-
-estimators = []
-estimators.append(("SGDRegressor", SGDRegressor))
-estimators.append(("KNeighborsRegressor", KNeighborsRegressor))
-estimators.append(("SVR", SVR))
-estimators.append(("RandomForestRegressor", RandomForestRegressor))
-
-train_sizes = np.geomspace(1e-3, 1.0, 8)
-fig = plot_learning_curve_times(estimators, X, y, train_sizes, cv=cv)
-fig.show()
